Project plan save and activate post API changes done
