+++ src/components/Projects/ProjectPlan/ProjectPlanList.js
@@ -1,12 +1,10 @@
 import React, { useState, useEffect, useRef, Suspense } from "react";
-import { useDispatch, useSelector } from "react-redux";
 import { TreeTable } from "primereact/treetable";
 import { Column } from "primereact/column";
 import { ProjectService } from "../../../service/PegaService";
 import ConfirmationPopUp from "../ConfirmationPopUp";
 import filter from "../../../assets/images/filter.svg";
 import { changeDateFormat } from "../../../utils";
-// import { projectPlan } from "../../../store/actions/ProjectActions";
 import BlueFilter from "../../../assets/images/BlueFilterIcon.svg";
 import complete from "../../../assets/images/complete.svg";
 import hyphen from "../../../assets/images/hyphen.svg";
@@ -16,22 +14,14 @@ import Awaiting from "../../../assets/images/Awaiting.svg";
 import override from "../../../assets/images/override.svg";
 import { Dropdown } from "primereact/dropdown";
 import { useNavigate } from "react-router-dom";
-// import { selectedProject } from "../../../store/actions/ProjectSetupActions";
 import { InputNumber } from "primereact/inputnumber";
-import { getProjectPlan } from "../../../apis/projectPlanApi";
 import "./index.scss";
 import TaskDialog from "../../TaskDialog";
-import {
-  updateProjectPlanAction,
-  updateProjectPlanDesignAction,
-} from "../../../store/actions/ProjectPlanActions";
-import moment from "moment";
 import ApproveDesignDialog from "./ApproveDesignDialog";
 import { useLocation } from "react-router-dom";
 import CPPFA from "./../../AWMJobs/CPPFA";
 
-const ProjectPlanList = (props) => {
-  const [pegadata, setPegaData] = useState(null);
+const ProjectPlanList = ({ projectPlan, selectedProject, projectPlanDesign, pegadata, setPegaData, setUpdatedProjectPlanDesignData, setActiveSave }) => {
   const [ProjectFrozen, setProjectFrozen] = useState(false);
   const [frozenCoulmns, setFrozenColumn] = useState([]);
   const [selectedColumnName, setSelectedColumnName] = useState(null);
@@ -47,15 +37,7 @@ const ProjectPlanList = (props) => {
   );
   const [flag, setFlag] = useState("");
   const [loader, setLoader] = useState(false);
-  const projectSetup = useSelector((state) => state.ProjectSetupReducer);
-  const selectedProjectDetails = projectSetup.selectedProject;
-  const mode = projectSetup.mode;
-  // console.log("selectedProjectDetails: ", selectedProjectDetails);
-  const projectPlanReducer = useSelector((state) => state.ProjectPlanReducer);
-  const ProjectPlanData = projectPlanReducer.projectPlan;
-  const { loading } = projectPlanReducer;
-
-  const dispatch = useDispatch();
+  //projectPlanDesign
   const navigate = useNavigate();
 
   const op = useRef(null);
@@ -76,209 +58,7 @@ const ProjectPlanList = (props) => {
 
   useEffect(() => {
     (async () => {
-      let restructuredData = [];
-      setLoader(true);
-      const apiData =
-        mode === "design" && selectedProjectDetails.Project_ID
-          ? await getProjectPlan(selectedProjectDetails.Project_ID)
-          : [];
-      setLoader(false);
-      console.log("projectPlanApiData::", apiData);
-      apiData && dispatch(updateProjectPlanDesignAction(apiData));
-      restructuredData =
-        apiData?.length > 0 ? getRestructuredData(apiData) : [];
-      dispatch(updateProjectPlanAction(restructuredData));
-    })();
-  }, [mode]);
-
-  const getRestructuredData = (apiData) => {
-    let mainTempArr = [];
-
-    const tasks = [
-      {
-        name: "Define Design Intent",
-        code: "DDI",
-        data: apiData.filter((data) => data.AWM_Task_ID.includes("DDI_")),
-      },
-      {
-        name: "Upload Approved Design Intent",
-        code: "UADI",
-        data: apiData.filter((data) => data.AWM_Task_ID.includes("UADI_")),
-      },
-      {
-        name: "Define Design Template",
-        code: "DDT",
-        data: apiData.filter((data) => data.AWM_Task_ID.includes("DDT_")),
-      },
-      {
-        name: "Upload Regional Design Template",
-        code: "URDT",
-        data: apiData.filter((data) => data.AWM_Task_ID.includes("URDT_")),
-      },
-      {
-        name: "Approve Regional Design Template",
-        code: "ARDT",
-        data: apiData.filter((data) => data.AWM_Task_ID.includes("ARDT_")),
-      },
-      {
-        name: "Define Production Ready Art",
-        code: "DPRA",
-        data: apiData.filter((data) => data.AWM_Task_ID.includes("DPRA_")),
-      },
-      {
-        name: "Upload Production Ready Art",
-        code: "UPRA",
-        data: apiData.filter((data) => data.AWM_Task_ID.includes("UPRA_")),
-      },
-      {
-        name: "Approve Production Ready Art",
-        code: "APRA",
-        data: apiData.filter((data) => data.AWM_Task_ID.includes("APRA_")),
-      },
-
-      {
-        name: "Confirm Preliminary print feasibility Assessment done (& upload documents - optional)",
-        code: "CPPFA",
-        data: apiData.filter((data) => data.AWM_Task_ID.includes("CPPFA_")),
-      },
-      {
-        name: "Define New Print Feasibility Scope",
-        code: "DNPF",
-        data: apiData.filter((data) => data.AWM_Task_ID.includes("DNPF_")),
-      },
-      {
-        name: "Color Confirm Development done (& upload documents - optional) (can be multiple)",
-        code: "CCD",
-        data: apiData.filter((data) => data.AWM_Task_ID.includes("CCD_")),
-      },
-      {
-        name: "Confirm Print Trial (if applicable) done (& upload documents - optional) (can be multiple)",
-        code: "CPT",
-        data: apiData.filter((data) => data.AWM_Task_ID.includes("CPT_")),
-      },
-      {
-        name: "Define New Link Ink Qualification scope",
-        code: "DNIQ",
-        data: apiData.filter((data) => data.AWM_Task_ID.includes("DNIQ_")),
-      },
-      {
-        name: "Confirm New Ink Qualification done (& upload documents - optional) (can be multiple)",
-        code: "CNIQ",
-        data: apiData.filter((data) => data.AWM_Task_ID.includes("CNIQ_")),
-      },
-    ];
-    // console.log("tasks:", tasks);
-    tasks.forEach((task) => {
-      // console.log("taskLength", task.data?.length);
-      if (task.data?.length === 1) {
-        let tempObj = {};
-        tempObj["key"] = task.data[0].AWM_Task_ID;
-
-        let dataObj = {};
-        dataObj["Task"] = task.data[0].Task_Name;
-        dataObj["Dependency"] = task.data[0].Dependency;
-        dataObj["Role"] = task.data[0].Role;
-        dataObj["RoleOptions"] = task.data[0].RoleOptions;
-        dataObj["Assignee"] = task.data[0].Assignee;
-        dataObj["OwnerOptions"] = task.data[0].OwnerOptions;
-        dataObj["State"] = task.data[0].State;
-        dataObj["Duration"] = task.data[0].Duration;
-        dataObj["StartDate"] = task.data[0].Start_Date;
-        dataObj["EndDate"] = task.data[0].End_Date;
-        dataObj["ConsumedBuffer"] = task.data[0].Consumed_Buffer;
-        dataObj["HelpNeeded"] = task.data[0].Help_Needed;
-
-        tempObj["data"] = dataObj;
-        tempObj["children"] = [];
-        tempObj["redirect"] = true;
-
-        mainTempArr.push(tempObj);
-      } else if (task.data?.length > 1) {
-        let tempObj = {};
-        let tempArr = [];
-        let pStartDate = "";
-        let pEndDate = "";
-        let startDateArr = [];
-        let endDateArr = [];
-        // let startDateArr = [
-        //   "20230411T000000.000 GMT",
-        //   "20230410T000000.000 GMT",
-        // ];
-        // let endDateArr = ["20230412T000000.000 GMT", "20230411T000000.000 GMT"];
-
-        tempObj["key"] = task.code;
-        let dataObj = {};
-        dataObj["Task"] = `${task.name} (X${task.data?.length})`;
-        dataObj["Dependency"] = task.data[0].Dependency;
-        dataObj["Role"] = "";
-        dataObj["RoleOptions"] = "";
-        dataObj["Assignee"] = "";
-        dataObj["OwnerOptions"] = "";
-        dataObj["State"] = "";
-        dataObj["Duration"] = "";
-
-        dataObj["ConsumedBuffer"] = "";
-        dataObj["HelpNeeded"] = false;
-
-        tempObj["data"] = dataObj;
-        tempObj["redirect"] = true;
-
-        //child array creation
-
-        task.data.forEach((dt, index) => {
-          dt.Start_Date && startDateArr?.push(dt.Start_Date);
-          dt.End_Date && endDateArr?.push(dt.End_Date);
-          pStartDate =
-            startDateArr.length > 0 &&
-            moment.min(
-              startDateArr.map((date) =>
-                moment(date, "YYYYMMDDTHHmmss.SSS [GMT]")
-              )
-            );
-          pEndDate =
-            endDateArr.length > 0 &&
-            moment.max(
-              endDateArr.map((date) =>
-                moment(date, "YYYYMMDDTHHmmss.SSS [GMT]")
-              )
-            );
-          let tempObj = {};
-          tempObj["key"] = dt.AWM_Task_ID;
-
-          let dataObj = {};
-          dataObj["Task"] = `${index + 1}). ${dt.Task_Name}`;
-          dataObj["Dependency"] = dt.Dependency;
-          dataObj["Role"] = dt.Role;
-          dataObj["RoleOptions"] = dt.RoleOptions;
-          dataObj["Assignee"] = dt.Assignee;
-          dataObj["OwnerOptions"] = dt.OwnerOptions;
-          dataObj["State"] = dt.State;
-          dataObj["Duration"] = dt.Duration;
-          dataObj["StartDate"] = dt.Start_Date;
-          dataObj["EndDate"] = dt.End_Date;
-          dataObj["ConsumedBuffer"] = dt.Consumed_Buffer;
-          dataObj["HelpNeeded"] = dt.Help_Needed;
-
-          tempObj["data"] = dataObj;
-          tempObj["children"] = [];
-
-          tempArr.push(tempObj);
-        });
-        dataObj["StartDate"] = pStartDate;
-        dataObj["EndDate"] = pEndDate;
-
-        tempObj["children"] = tempArr;
-
-        mainTempArr.push(tempObj);
-      }
-    });
-    return mainTempArr; //toBeReplacedWithapiData;
-  };
-
-  useEffect(() => {
-    (async () => {
-      try {
-        const ProjectPlanData = projectPlanReducer.projectPlan;
+      try {      
 
         let filteredPegaDataJson = localStorage.getItem("columnWiseFilterData");
         const filteredPegaData = JSON.parse(filteredPegaDataJson);
@@ -286,8 +66,10 @@ const ProjectPlanList = (props) => {
         if (filteredPegaData && filteredPegaData.length) {
           setFilters(filteredPegaData);
           setSelectedCities(filteredPegaData);
-          setPegaData(ProjectPlanData);
-        } else setPegaData(ProjectPlanData);
+          setPegaData(projectPlan);
+        } else {
+          setPegaData(projectPlan);
+        };
         // according to pathname we need to call api and store column name in local storage
         let columnNamesJson = localStorage.getItem("projectPlanAllColumnNames");
         const columnNames = JSON.parse(columnNamesJson);
@@ -307,7 +89,7 @@ const ProjectPlanList = (props) => {
         const sortingData = JSON.parse(jsonSortingData1);
 
         if (sortingData && sortingData.length) {
-          const sortedData = [...ProjectPlanData].sort((a, b) => {
+          const sortedData = [...projectPlan].sort((a, b) => {
             return a[sortingData[0]] > b[sortingData[0]] ? 1 : -1;
           });
 
@@ -329,7 +111,7 @@ const ProjectPlanList = (props) => {
       }
     })();
     // setLoading(false);
-  }, [projectPlanReducer.projectPlan]);
+  }, [projectPlan]);
 
   const addFrozenColumns = (name) => {
     if (!frozenCoulmns.includes(name)) {
@@ -444,7 +226,7 @@ const ProjectPlanList = (props) => {
     const field = rowData.field;
     const optionsData = options.data;
     const currentUrl = location.pathname;
-    let currentUrlLastSeg = currentUrl.split("/")[2];
+    let currentUrlLastSeg = currentUrl.split("/")[3];
     const key = options?.key;
     const keyCode = key?.split("_");
     const url = `MyTasks/${keyCode[0]}/${key}/${currentUrlLastSeg}`;
@@ -460,7 +242,6 @@ const ProjectPlanList = (props) => {
             }`}
             onClick={() => {
               if (field && field.length && keyCode[0] !== "CPPFA") {
-                // dispatch(selectedProject(optionsData, "My Projects"));
                 (options.redirect === true || optionsData.Task) &&
                   navigate(`../${url}`, { replace: true });
               } else {
@@ -629,7 +410,7 @@ const ProjectPlanList = (props) => {
     pegadata.forEach((data) => {
       if (data.children.length === 0) {
         let obj = {};
-        obj["AWM_Project_ID"] = selectedProjectDetails.Project_ID; //to be changed
+        obj["AWM_Project_ID"] = selectedProject.Project_ID; //to be changed
         obj["Help_Needed"] = data.data.HelpNeeded;
         obj["AWM_Task_ID"] = data.key;
         obj["End_Date"] = data.data.EndDate;
@@ -645,7 +426,7 @@ const ProjectPlanList = (props) => {
       } else if (data.children.length > 0) {
         data.children.forEach((child) => {
           let tempObj = {};
-          tempObj["AWM_Project_ID"] = selectedProjectDetails.Project_ID;
+          tempObj["AWM_Project_ID"] = selectedProject.Project_ID;
           tempObj["Help_Needed"] = child.data.HelpNeeded;
           tempObj["AWM_Task_ID"] = child.key;
           tempObj["End_Date"] = child.data.EndDate;
@@ -661,11 +442,10 @@ const ProjectPlanList = (props) => {
         });
       }
     });
-    dispatch(updateProjectPlanDesignAction(arr));
+    setUpdatedProjectPlanDesignData(arr);
   };
 
   useEffect(() => {
-    // alert(pegadata);
     pegadata && updateProjectPlanDesign();
   }, [pegadata]);
 
@@ -674,13 +454,15 @@ const ProjectPlanList = (props) => {
     rowData.data[ele] = value.Name;
     console.log("Pegadata: ", pegadata);
     setPegaData([...pegadata]);
-    updateProjectPlanDesign();
+    setActiveSave(false);
+   //updateProjectPlanDesign();
   };
 
   const onDurationChange = (rowData, { value }, ele) => {
     rowData.data[ele] = value < 1 ? "0" : value?.toString();
     console.log("Pegadata: ", pegadata);
     setPegaData([...pegadata]);
+    setActiveSave(false);
   };
 
   const rowExpansionColumns = () => {
